src/AI/version1.2/requirements.txt
absl-py==0.2.2
alabaster==0.7.11
astor==0.7.1
atomicwrites==1.1.5
attrs==18.1.0
Babel==2.6.0
backcall==0.1.0
bintrees==2.0.7
bleach==3.1.0
certifi==2018.4.16
click==6.7
colorama==0.4.1
colorlover==0.3.0
cufflinks==0.14.6
cycler==0.10.0
dash==0.36.0
dash-core-components==0.43.0
dash-daq==0.1.0
dash-html-components==0.13.5
dash-renderer==0.18.0
dash-table==3.1.11
decorator==4.3.0
docutils==0.14
entrypoints==0.3
Flask==0.12.2
Flask-Compress==1.4.0
Flask-Cors==3.0.3
future==0.17.1
gast==0.2.0
gdax==1.0.6
GPy==1.8.4
-e git+https://github.com/SheffieldML/GPyOpt.git@fff3f65c4e76bed69ce6dc55c459a5d37db5b243#egg=GPyOpt
grpcio==1.13.0
h5py==2.9.0
html5lib==0.9999999
hyperopt==0.1.1
imagesize==1.0.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
more-itertools==4.2.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
nose==1.3.7
notebook==5.7.0
numpy==1.13.3
packaging==17.1
pandas==0.20.3
pandocfilters==1.4.2
paramz==0.9.2
parso==0.3.1
pickleshare==0.7.5
Pillow==5.2.0
plotly==3.6.1
pluggy==0.6.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.0
psycopg2==2.7.3
py==1.5.4
Pygments==2.3.1
pymongo==3.7.2
pyparsing==2.2.0
pytest==3.6.3
python-box==3.1.1
python-dateutil==2.7.5
pytz==2018.5
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
requests==2.13.0
retrying==1.3.3
ruamel.yaml==0.15.87
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
slackclient==1.3.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.13
ta-lib==0.4.17
tensorboard==1.9.0
tensorflow-gpu==1.5.0rc1
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.40.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wincertstore==0.2
